fix: support Windows RUN heredoc fixes

+++ internal/integration/__snapshots__/TestFixRealWorldMetalama_1.snap.Dockerfile
@@ -21,12 +21,16 @@ ENV RUNNING_IN_DOCKER=TRUE
 ENV PATH="C:\Windows\System32\WindowsPowerShell\v1.0;${PATH}"
 
 # Enable long path support
-RUN Set-ItemProperty -Path 'HKLM:\SYSTEM\CurrentControlSet\Control\FileSystem' -Name 'LongPathsEnabled' -Value 1
-
-# Install Git
-RUN Invoke-WebRequest -Uri https://github.com/git-for-windows/git/releases/download/v2.50.0.windows.1/PortableGit-2.50.0-64-bit.7z.exe -OutFile PortableGit.exe; `
-    Start-Process -FilePath .\PortableGit.exe -ArgumentList '-o"C:\git"', '-y' -Wait; `
-    Remove-Item PortableGit.exe
+RUN <<EOF
+$ErrorActionPreference = 'Stop'
+Set-ItemProperty -Path 'HKLM:\SYSTEM\CurrentControlSet\Control\FileSystem' -Name 'LongPathsEnabled' -Value 1
+if (-not $?) { if ($LASTEXITCODE -ne 0) { exit $LASTEXITCODE }; exit 1 }
+Invoke-WebRequest -Uri https://github.com/git-for-windows/git/releases/download/v2.50.0.windows.1/PortableGit-2.50.0-64-bit.7z.exe -OutFile PortableGit.exe
+if (-not $?) { if ($LASTEXITCODE -ne 0) { exit $LASTEXITCODE }; exit 1 }
+Start-Process -FilePath .\PortableGit.exe -ArgumentList '-o"C:\git"', '-y' -Wait
+if (-not $?) { if ($LASTEXITCODE -ne 0) { exit $LASTEXITCODE }; exit 1 }
+Remove-Item PortableGit.exe
+EOF
 
 # Add git to PATH using ENV directive (persists across shell switches)
 ENV PATH="C:\git\cmd;C:\git\bin;C:\git\usr\bin;${PATH}"
@@ -59,13 +63,14 @@ RUN Invoke-WebRequest -Uri https://dot.net/v1/dotnet-install.ps1 -OutFile dotnet
 ENV PATH="C:\Program Files\dotnet;${PATH}"
 
 # Install .NET Sdk 8.0.414
-RUN & .\dotnet-install.ps1 -Version 8.0.414 -InstallDir 'C:\Program Files\dotnet'
-
-# Install .NET DotNetRuntime 9.0.9
-RUN & .\dotnet-install.ps1 -Version 9.0.9 -Runtime dotnet -InstallDir 'C:\Program Files\dotnet'
-
-# Install .NET Sdk 10.0.100
-RUN & .\dotnet-install.ps1 -Version 10.0.100 -InstallDir 'C:\Program Files\dotnet'
+RUN <<EOF
+$ErrorActionPreference = 'Stop'
+& .\dotnet-install.ps1 -Version 8.0.414 -InstallDir 'C:\Program Files\dotnet'
+if (-not $?) { if ($LASTEXITCODE -ne 0) { exit $LASTEXITCODE }; exit 1 }
+& .\dotnet-install.ps1 -Version 9.0.9 -Runtime dotnet -InstallDir 'C:\Program Files\dotnet'
+if (-not $?) { if ($LASTEXITCODE -ne 0) { exit $LASTEXITCODE }; exit 1 }
+& .\dotnet-install.ps1 -Version 10.0.100 -InstallDir 'C:\Program Files\dotnet'
+EOF
 
 # Install VS Build Tools
 COPY VisualStudio.17.14.15.Release.chman /VisualStudio.17.14.15.Release.chman
